Move mutation code for clusters into its own package
